% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{myMean}
\alias{myMean}
\title{Mean  function}
\usage{
myMean(x)
}
\arguments{
\item{txt}{list of \code{numeric vectors}}
}
\value{
vector of means
}
\description{
Mean  function
}
\details{
Calculates mean for each provided vector in list
}
\examples{
myMean(list(c(1,3)))
}
